+++ k2EngineLow/MiniEngine/raytracing/TLASBuffer.cpp
@@ -23,7 +23,7 @@ namespace raytracing {
 		uint64_t tlasSize;
 		auto d3dDevice = g_graphicsEngine->GetD3DDevice();
 
-		int numInstance = instances.size();
+		int numInstance = static_cast<int>(instances.size());
 		D3D12_BUILD_RAYTRACING_ACCELERATION_STRUCTURE_INPUTS inputs = {};
 		inputs.DescsLayout = D3D12_ELEMENTS_LAYOUT_ARRAY;
 		inputs.Flags = D3D12_RAYTRACING_ACCELERATION_STRUCTURE_BUILD_FLAG_ALLOW_UPDATE;
